[v2] Fix E_ACCESSDENIED error when renaming directory
There are cases when file managers or file explorers are used to access
and modify save data files in the save data directory. This may cause
the save data directory to be locked if those applications do not
properly release or close the directory handle. Thus, when calls to
`DirectorySaveDataFileSystem.DoCommit()` are made, which would invoke
`LocalFileSystem.RenameDirInternal()` in turn, .NET would keep throwing
an `IOException` with an `HResult` error code of 0x80070005
(`E_ACCESSDENIED`), indicating an Access Denied error.

This commit fixes this issue by performing either of the following:

1. On Windows, manually rename the `source` directory to the `dest`
directory via the P/Invoke API in `shobjidl_core.h`, which should be
used by Windows Explorer in theory.
2. On other Unix-based operating systems, manually call the `rename`
system call via the Mono framework to execute the renaming.

It seems that `DirectoryInfo` operations are known to be quite janky and
sensitive to those directories being used by other processes on the
machine. This proposed implementation is supposed to be equivalent to
the previous implementation and it is tested to be working on Ryujinx.
The operation remains atomic and hence, it should not leave the file
system in an inconsistent state.

Signed-off-by: James Raphael Tiovalen <jamestiotio@gmail.com>